requirements.txt
flask
open-interpreter